Merge

+++ admin/php-templates/midwife/view.php
@@ -5,16 +5,21 @@
 session_start();
 
 @include '../php-templates/redirect/admin-page-setter.php';
- 
+@include '../php-templates/redirect/not-for-patient.php';
 
-// fetch nurses 
-$select = "SELECT u.id AS id, 
+$midwife_select = "SELECT u.id AS id, 
   CONCAT(u.first_name,IF(u.mid_initial='', '', CONCAT(' ',u.mid_initial,'.')),' ',u.last_name) AS name, u.email,  
-  IF(u.status=0, 'Inactive', 'Active') AS status, d.contact_no, d.b_date,  health_center,
+  IF(u.status=0, 'Inactive', 'Active') AS status, d.contact_no, d.b_date,
   details_id 
-  FROM users as u, details as d,
-  barangay as b WHERE u.admin = 0 AND d.id=u.details_id AND d.barangay_id=b.id;";
-$result = mysqli_query($conn, $select);
+  FROM users as u, details as d 
+  WHERE u.admin = 0 AND d.id=u.details_id";
+// fetch nurses 
+$select = "SELECT main.id, name, email, main.status, contact_no, b_date, details_id, health_center
+FROM 
+	($midwife_select) as main 
+LEFT JOIN barangay as b
+ON main.id=b.assigned_midwife;";
+$result = mysqli_query($conn, $admin==1?$select:$midwife_select);
 $midwife_list = [];
 
 if(mysqli_num_rows($result))  {
@@ -25,7 +30,7 @@ if(mysqli_num_rows($result))  {
     $s = $row['status'];  
     $c_no = $row['contact_no'];  
     $b_date = $row['b_date'];  
-    $bgy = $row['health_center'];  
+    $bgy = $admin==1?$row['health_center']:'';  
     $det_id = $row['details_id'];  
     array_push($midwife_list, array(
       'id' => $id,
@@ -58,14 +63,7 @@ include_once('../php-templates/admin-navigation-head.php');
 
   <!-- Sidebar -->
  
-  <?php include_once('../php-templates/admin-navigation-left.php'); 
-  
-    // from edit.php 
-    if (isset($_GET['edit']) && $_GET['edit']==0) {
-      echo "<script>alert(\"Can't edit this account!\");</script>";
-    }   
-
-  ?>
+  <?php include_once('../php-templates/admin-navigation-left.php'); ?>
  
 <!-- style css -->
 <style>
@@ -113,15 +111,16 @@ include_once('../php-templates/admin-navigation-head.php');
                   <!-- <th scope="col">Birthdate</th> -->
                   <th scope="col">Barangay</th>
                 <?php } ?>
- 
-                <th scope="col">Actions</th>
+                <?php if ($admin==1) {?> 
+                  <th scope="col">Actions</th>
+                <?php }?>
               </tr>
             </thead>
             <tbody>
               <?php 
                 foreach ($midwife_list as $key => $value) {
               ?>    
-              <?php  ?>
+              
                 <tr>
                   <th scope="row"><?php echo $key+1; ?></th>
                   <td><?php echo $value['name']; ?></td>
@@ -134,14 +133,15 @@ include_once('../php-templates/admin-navigation-head.php');
                     <td><?php echo $value['contact']; ?></td>
                     <!-- <td><?php //$dtf = date_create($value['b_date']); echo date_format($dtf,"F d, Y"); ?></td> -->
                     <td><?php echo $value['barangay']; ?></td>
-                  <?php } ?>
-                  <td>
-                    <button class="edit btn btn-success btn-sm btn-inverse"><a href="edit-midwife.php?id=<?php echo $value['id'] ?>">Edit</a></button>
-                    <?php if ($value['id']!=$_SESSION['id']) { ?>
-                      <button class="del btn btn-danger btn-sm btn-inverse"><a href="delete-midwife.php?id=<?php echo $value['id'] ?>&details_id=<?php echo $value['details_id'] ?>">Delete</a></button> 
-                    <?php } ?>
- 
-                  </td>
+                  <?php } ?> 
+                  <?php if ($admin==1) {?>
+                    <td>
+                      <button class="edit btn btn-success btn-sm btn-inverse"><a href="edit-midwife.php?id=<?php echo $value['id'] ?>">Edit</a></button>
+                      <?php if ($value['id']!=$_SESSION['id']) { ?>
+                        <button class="del btn btn-danger btn-sm btn-inverse"><a href="delete-midwife.php?id=<?php echo $value['id'] ?>&details_id=<?php echo $value['details_id'] ?>">Delete</a></button> 
+                      <?php } ?> 
+                    </td>
+                  <?php }?> 
                 </tr>
               <?php 
                 }
